pyproject.toml
[project]
name = "filmitorrent-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "boltons>=25.0.0",
    "desktop-notifier>=6.2.0",
    "dropbox>=12.0.2",
    "hotpdf>=0.5.2",
    "icecream>=2.1.5",
    "jurigged>=0.6.1",
    "logging>=0.4.9.6",
    "loguru>=0.7.3",
    "orjson>=3.11.2",
    "pdfkit>=1.0.0",
    "pillow>=11.3.0",
    "playwright>=1.54.0",
    "pydoll-python>=2.6.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "setuptools>=80.9.0",
    "simplejson>=3.20.1",
    "tqdm>=4.67.1",
    "ujson>=5.10.0",
    "urllib3>=2.5.0",
    "wheel>=0.45.1",
]
